added some 'double' to reduce picky warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@5130 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/MathTypes.cpp
@@ -574,7 +574,7 @@ void FQuat::GetMatrix(FMatrix3 &mat) const
  * Spherical Linear Interpolation.
  * As f goes from 0 to 1, the quaternion goes from "from" to "to".
  */
-void FQuat::Slerp(const FQuat &from, const FQuat &to, float f)
+void FQuat::Slerp(const FQuat &from, const FQuat &to, double f)
 {
 	/// Reference: Shoemake at SIGGRAPH 89, See also:
 	/// http://www.gamasutra.com/features/19980703/quaternions_01.htm
